Tests decomposition. Add utils for tests.
